# Optimize normalize_name